pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.116.2",
    "google-adk>=1.14.1",
    "httpx>=0.28.1",
]
